Previous/requirement.txt
librosa
matplotlib
numpy
pyyaml
scikit-learn
tqdm
streamlit